chore: cleanup, truecolor, and toml
