+++ go.mod
@@ -2,10 +2,4 @@ module github.com/mikesmithgh/bgps
 
 go 1.22.0
 
-require github.com/fatih/color v1.16.0
-
-require (
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	golang.org/x/sys v0.14.0 // indirect
-)
+require github.com/pelletier/go-toml/v2 v2.1.1
